+++ src/main/kotlin/io/Paths.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-@file:JvmName("FileUtils")
+@file:JvmName("KFiles")
 @file:Suppress("NOTHING_TO_INLINE")
 
 package moe.kanon.kommons.io